% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{build_preamble}
\alias{build_preamble}
\title{Build LaTeX preamble}
\usage{
build_preamble(doc)
}
\arguments{
\item{doc}{LaTeX document string.}
}
\value{
A character vector of LaTeX preamble lines.
}
\description{
Generate the LaTeX preamble (packages, color definitions, environments) needed
for the knitted output. Only includes items that are not already defined in
the document.
}
\keyword{internal}
