% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{check_color_definition}
\alias{check_color_definition}
\title{Check color definition in preamble}
\usage{
check_color_definition(clean_preamble)
}
\arguments{
\item{clean_preamble}{Document preamble with LaTeX comments stripped.}
}
\description{
Validate that \verb{\\\\definecolor\{...\}} in the document preamble is preceded by
\verb{\\\\usepackage\{xcolor\}}. Raises an error if xcolor is missing.
}
\keyword{internal}
