% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{chunk_counter}
\alias{chunk_counter}
\title{Chunk counter}
\usage{
chunk_counter(reset = FALSE)
}
\arguments{
\item{reset}{If \code{TRUE}, reset the counter to 0.}
}
\value{
The current chunk number (incremented after each call).
}
\description{
A closure that tracks the number of chunks processed. Resets on each
top-level \code{knit()} call.
}
\keyword{internal}
