% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{detect_bib_engine}
\alias{detect_bib_engine}
\title{Detect bibliography engine}
\usage{
detect_bib_engine(tex_content)
}
\arguments{
\item{tex_content}{LaTeX document as a single string.}
}
\value{
\code{"biber"}, \code{"bibtex"}, or \code{NULL}.
}
\description{
Detect which bibliography engine is needed based on LaTeX document content.
}
\keyword{internal}
