% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{digest}
\alias{digest}
\title{Compute a digest hash}
\usage{
digest(x)
}
\arguments{
\item{x}{Any R object to hash.}
}
\value{
A character string of the hash.
}
\description{
Compute an MD5 (or SHA1 if \pkg{digest} is not available) hash of an R object.
Used to generate cache keys.
}
\keyword{internal}
