% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{doc_bib_files}
\alias{doc_bib_files}
\title{Bibliography files a LaTeX document uses}
\usage{
doc_bib_files(text, dir)
}
\arguments{
\item{text}{The document text.}

\item{dir}{Folder the file names are relative to.}
}
\value{
Absolute paths of the files given to \code{\\addbibresource},
\code{\\addglobalbib} (and similar) and \code{\\bibliography}.
}
\description{
Bibliography files a LaTeX document uses
}
\keyword{internal}
