% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{eval_lang}
\alias{eval_lang}
\title{Evaluate a language object}
\usage{
eval_lang(x, envir = knit_global())
}
\arguments{
\item{x}{An R language object (expression, call, or name).}

\item{envir}{Environment in which to evaluate.}
}
\value{
Evaluated result or the original object on error.
}
\description{
Evaluate an expression, call, or symbol in a given environment. Returns
the object unchanged if evaluation fails.
}
\keyword{internal}
