% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{find_up}
\alias{find_up}
\title{Nearest folder at or above a path that contains a file}
\usage{
find_up(path, name)
}
\arguments{
\item{path}{Starting folder.}

\item{name}{File or folder name to look for.}
}
\value{
The folder, or \code{NULL}.
}
\description{
Nearest folder at or above a path that contains a file
}
\keyword{internal}
