% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{has_defineverbatimenvironment}
\alias{has_defineverbatimenvironment}
\title{Check if Verbatim environment is defined}
\usage{
has_defineverbatimenvironment(doc, env)
}
\description{
Check if Verbatim environment is defined
}
\keyword{internal}
