% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{has_package}
\alias{has_package}
\title{Check if LaTeX package is loaded}
\usage{
has_package(doc, pkg)
}
\arguments{
\item{doc}{LaTeX document string.}

\item{pkg}{Package name to check.}
}
\value{
\code{TRUE} if the package is loaded in the document preamble.
}
\description{
Check if LaTeX package is loaded
}
\keyword{internal}
