% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{hilight_source}
\alias{hilight_source}
\title{Highlight R source code}
\usage{
hilight_source(x, format = "latex")
}
\arguments{
\item{x}{Character vector of R code lines.}

\item{format}{Output format (\code{"latex"} or other).}
}
\value{
Character vector of highlighted code.
}
\description{
Apply syntax highlighting to R source code using the \pkg{highr} package.
}
\keyword{internal}
