% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{in_input_dir}
\alias{in_input_dir}
\title{Evaluate expression in input directory}
\usage{
in_input_dir(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
Result of the expression.
}
\description{
Temporarily change the working directory to the input file's directory
and evaluate an expression.
}
\keyword{internal}
