% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{insert_header}
\alias{insert_header}
\title{Insert preamble into document}
\usage{
insert_header(doc)
}
\arguments{
\item{doc}{LaTeX document string.}
}
\value{
Document string with preamble inserted.
}
\description{
Insert the generated LaTeX preamble (packages, color definitions, environments)
before \verb{\\\\begin\{document\}} in the output.
}
\keyword{internal}
