% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{knit_child}
\alias{knit_child}
\title{Knit a child document}
\usage{
knit_child(child)
}
\arguments{
\item{child}{Path to the child \code{.Rnw} file.}
}
\value{
Character string of LaTeX output from the child document.
}
\description{
Process a child \code{.Rnw} document and return its LaTeX output. Used when
a chunk has the \code{child} option set.
}
