% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{data}
\name{knit_code}
\alias{knit_code}
\title{Knit code store}
\format{
A list with methods: \code{get}, \code{set}, \code{delete}, \code{append},
\code{merge}, and \code{restore}.
}
\usage{
knit_code
}
\description{
A list-like object (created by \code{\link[=new_defaults]{new_defaults()}}) that stores the code for each
chunk by label. Populated during parsing and consumed during execution.
}
\keyword{datasets}
