% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{knit_global}
\alias{knit_global}
\title{Get the knit evaluation environment}
\usage{
knit_global()
}
\value{
An environment object.
}
\description{
Returns the environment in which code chunks are evaluated. Set by \code{\link[=knit]{knit()}}
via \code{knit_global_set()}.
}
\keyword{internal}
