% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{knit_hooks}
\alias{knit_hooks}
\title{Knit output hooks}
\format{
A list with methods: \code{get}, \code{set}, \code{delete}, \code{append},
\code{merge}, and \code{restore}. Default hooks include:
\describe{
\item{source}{Format source code display}
\item{output}{Format text output}
\item{warning}{Format warnings}
\item{message}{Format messages}
\item{error}{Format errors}
\item{plot}{Format plots}
\item{inline}{Format inline results}
\item{chunk}{Wrap entire chunk output}
\item{evaluate}{Function used to evaluate code}
\item{evaluate.inline}{Function used to evaluate inline expressions}
}
}
\usage{
knit_hooks
}
\description{
A list-like object (created by \code{\link[=new_defaults]{new_defaults()}}) that stores output hook functions.
Hooks control how different types of output (source code, text output, warnings,
messages, errors, plots, inline results) are rendered in the final document.
}
\seealso{
\code{\link{render_latex}} to install LaTeX-specific hooks
}
\keyword{datasets}
