% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\docType{data}
\name{knit_patterns}
\alias{knit_patterns}
\title{Knit patterns}
\format{
A list with methods: \code{get}, \code{set}, \code{delete}, \code{append},
\code{merge}, and \code{restore}.
}
\usage{
knit_patterns
}
\description{
A list-like object (created by \code{\link[=new_defaults]{new_defaults()}}) that stores the active set of
pattern definitions used to parse the input document. Defaults to the \code{rnw}
pattern list from \code{\link[=all_patterns]{all_patterns()}}.
}
\seealso{
\code{\link{all_patterns}}
}
\keyword{datasets}
