% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{merge_list}
\alias{merge_list}
\title{Merge two named lists}
\usage{
merge_list(x, y)
}
\arguments{
\item{x}{Base list}

\item{y}{List whose values override \code{x}}
}
\value{
A merged list.
}
\description{
Merge two named lists
}
\keyword{internal}
