% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{new_defaults}
\alias{new_defaults}
\title{Create a new defaults container}
\usage{
new_defaults(value = list())
}
\arguments{
\item{value}{A named list of initial default values.}
}
\value{
A list of functions for manipulating the defaults.
}
\description{
Creates an object with \code{get}, \code{set}, \code{delete}, \code{append}, \code{merge}, and \code{restore}
methods for managing a list of default values. Used internally by \code{opts_chunk},
\code{opts_knit}, \code{knit_hooks}, \code{knit_patterns}, and \code{knit_code}.
}
\keyword{internal}
