% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%n\%}
\title{Null-default operator}
\usage{
x \%n\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Fallback value}
}
\value{
\code{x} if non-NULL, else \code{y}.
}
\description{
Returns \code{y} if \code{x} is \code{NULL}, otherwise returns \code{x}.
Inspired by the \verb{\\\%||\\\%} operator from \pkg{knitr}.
}
\keyword{internal}
