% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_chunk}
\alias{opts_chunk}
\title{Default chunk options}
\format{
A list with methods: \code{get}, \code{set}, \code{delete}, \code{append},
\code{merge}, and \code{restore}.
}
\usage{
opts_chunk
}
\description{
A list-like object (created by \code{\link[=new_defaults]{new_defaults()}}) that stores default chunk options.
Use \code{opts_chunk$set(...)} to change defaults and \code{opts_chunk$get()} to retrieve them.
}
\examples{
opts_chunk$set(echo = FALSE, eval = TRUE)
opts_chunk$get("echo")
}
\seealso{
\code{\link{opts_knit}} for global options, \code{\link{knit_hooks}} for output hooks
}
\keyword{datasets}
