% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_knit}
\alias{opts_knit}
\title{Global knit options}
\format{
A list with methods: \code{get}, \code{set}, \code{delete}, \code{append},
\code{merge}, and \code{restore}.
}
\usage{
opts_knit
}
\description{
A list-like object (created by \code{\link[=new_defaults]{new_defaults()}}) that stores global knitrmini options.
Use \code{opts_knit$set(...)} to change options and \code{opts_knit$get()} to retrieve them.
}
\examples{
opts_knit$set(engine = "xelatex", progress = FALSE)
opts_knit$get("engine")
}
\seealso{
\code{\link{opts_chunk}} for chunk defaults
}
\keyword{datasets}
