% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{out_format}
\alias{out_format}
\title{Check output format}
\usage{
out_format(fmt = NULL)
}
\arguments{
\item{fmt}{Optional format name to test.}
}
\value{
Character string or logical.
}
\description{
Get or test the current output format. If called with no arguments, returns
the current format name. If called with a format name, returns \code{TRUE}
if the current format matches.
}
\keyword{internal}
