% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{packages_in_exprs}
\alias{packages_in_exprs}
\title{Package names used in parsed R code}
\usage{
packages_in_exprs(exprs)
}
\arguments{
\item{exprs}{An expression vector (from \code{parse()}), or \code{NULL}.}
}
\value{
Package names, possibly repeated or invalid (filtered later).
}
\description{
Package names used in parsed R code
}
\keyword{internal}
