% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\docType{data}
\name{pkg_bib_marker}
\alias{pkg_bib_marker}
\title{First line of every file written by \code{\link[=write_pkg_bib]{write_pkg_bib()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
pkg_bib_marker
}
\description{
First line of every file written by \code{\link[=write_pkg_bib]{write_pkg_bib()}}
}
\keyword{internal}
