% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{pkg_desc_sig}
\alias{pkg_desc_sig}
\title{Short checksums of installed packages' DESCRIPTION files}
\usage{
pkg_desc_sig(pkgs, lib = NULL)
}
\arguments{
\item{pkgs}{Package names.}

\item{lib}{Library paths (\code{NULL} for the session's).}
}
\value{
First 8 characters of each DESCRIPTION's md5 (\code{NA} if missing).
}
\description{
Short checksums of installed packages' DESCRIPTION files
}
\keyword{internal}
