% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{pkg_version}
\alias{pkg_version}
\title{Installed versions of packages}
\usage{
pkg_version(pkgs, lib = NULL)
}
\arguments{
\item{pkgs}{Package names.}

\item{lib}{Library paths (\code{NULL} for the session's).}
}
\value{
Character vector of versions.
}
\description{
Installed versions of packages
}
\keyword{internal}
