% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{project_files}
\alias{project_files}
\title{R, Rnw, Rmd and qmd files of the project containing a folder}
\usage{
project_files(path)
}
\arguments{
\item{path}{Folder in the project.}
}
\value{
Absolute paths.
}
\description{
R, Rnw, Rmd and qmd files of the project containing a folder
}
\keyword{internal}
