% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{r_code_blocks}
\alias{r_code_blocks}
\title{Blocks of R code in a file}
\usage{
r_code_blocks(file)
}
\arguments{
\item{file}{Path to an \code{.R}, \code{.Rnw}, \code{.Rmd} or \code{.qmd}.}
}
\value{
A list of character vectors: the whole file for \code{.R}, one
element per R chunk otherwise.
}
\description{
Blocks of R code in a file
}
\keyword{internal}
