% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{renv_dirs}
\alias{renv_dirs}
\title{Names of renv project folders}
\usage{
renv_dirs()
}
\value{
\code{RENV_PATHS_RENV} (if set and relative), \code{renv} and
\code{.renv}.
}
\description{
Names of renv project folders
}
\keyword{internal}
