% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{resolve_includegraphics}
\alias{resolve_includegraphics}
\title{Resolve includegraphics paths}
\usage{
resolve_includegraphics(doc, input_dir)
}
\arguments{
\item{doc}{LaTeX document string.}

\item{input_dir}{The directory of the input file.}
}
\value{
Document string with paths resolved.
}
\description{
Resolve relative paths in \verb{\\\\includegraphics} commands to absolute paths
rooted at \code{input_dir}.
}
\keyword{internal}
