% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{restore_verbatim}
\alias{restore_verbatim}
\title{Restore LaTeX verbatim constructs}
\usage{
restore_verbatim(input, original)
}
\arguments{
\item{input}{A single character string.}

\item{original}{A named character vector mapping placeholder to expression.}
}
\value{
The text with verbatim content restored.
}
\description{
Replace sentinel placeholders created by \code{\link[=protect_verbatim]{protect_verbatim()}} with their
original verbatim expressions.
}
\keyword{internal}
