% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{set_header}
\alias{set_header}
\title{Set LaTeX header content}
\usage{
set_header(...)
}
\arguments{
\item{...}{Named arguments of LaTeX code to add to the header.}
}
\description{
Add LaTeX code to the document preamble. Merges the provided values with
existing header entries stored in \code{opts_knit$get("header")}.
}
