% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{strip_latex_comments}
\alias{strip_latex_comments}
\title{Strip LaTeX comments}
\usage{
strip_latex_comments(doc)
}
\arguments{
\item{doc}{LaTeX document string.}
}
\value{
Document string with comments removed.
}
\description{
Remove LaTeX comment lines (starting with \code{\%}) from a document,
respecting escaped percent signs.
}
\keyword{internal}
