% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{unnamed_chunk}
\alias{unnamed_chunk}
\title{Generate unnamed chunk label}
\usage{
unnamed_chunk(prefix = NULL)
}
\arguments{
\item{prefix}{Label prefix (default from \code{opts_knit$get("unnamed.chunk.label")}).}
}
\value{
A unique chunk label string.
}
\description{
Generate unnamed chunk label
}
\keyword{internal}
