% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_inline}
\alias{parse_inline}
\title{Parse inline expressions}
\usage{
parse_inline(input, patterns)
}
\arguments{
\item{input}{Character vector of text lines.}

\item{patterns}{Pattern list (from \link{knit_patterns}).}
}
\value{
An \code{inline} object with fields \code{input}, \code{location}, and \code{code}.
}
\description{
Extract inline code expressions (e.g. \verb{\\\\Sexpr\{...\}}) from a block of text.
}
\keyword{internal}
